+++ libraries/ui-library/src/components/six-sidebar-item-group/six-sidebar-item-group.tsx
@@ -1,5 +1,6 @@
 import { Component, Element, h, Prop, State } from '@stencil/core';
 import { hasSlot } from '../../utils/slot';
+import { SixDetailsCustomEvent } from '../../components';
 
 /**
  * @since 1.0
@@ -61,6 +62,16 @@ export class SixSidebarItemGroup {
     return this.href != null && !this.hasItems;
   }
 
+  private handleDetailsShow = (event: SixDetailsCustomEvent<undefined>) => {
+    event.stopPropagation();
+    this.open = true;
+  };
+
+  private handleDetailsHide = (event: SixDetailsCustomEvent<undefined>) => {
+    event.stopPropagation();
+    this.open = false;
+  };
+
   private provideSlot = (name: string) => {
     if (this.summaryIconHasContent) {
       return (
@@ -94,12 +105,8 @@ export class SixSidebarItemGroup {
         inline={true}
         open={this.open}
         summary-icon={this.summaryIcon}
-        onSix-details-show={() => {
-          this.open = true;
-        }}
-        onSix-details-hide={() => {
-          this.open = false;
-        }}
+        onSix-details-show={this.handleDetailsShow}
+        onSix-details-hide={this.handleDetailsHide}
         hasContent={this.hasItems}
         tabindex={this.renderAsHref() ? -1 : undefined}
       >